move dislikedArtworksConnection off user and onto me schema
