+++ src/schema/v2/me/__tests__/index.test.ts
@@ -1,5 +1,6 @@
 /* eslint-disable promise/always-return */
 import gql from "lib/gql"
+import { HTTPError } from "lib/HTTPError"
 import { runAuthenticatedQuery, runQuery } from "schema/v2/test/utils"
 describe("me/index", () => {
   const query = gql`
@@ -306,6 +307,104 @@ describe("me/index", () => {
       }).then((data) => {
         expect(data).toEqual({ me: { emailConfirmed: false } })
       })
+    })
+  })
+
+  describe("dislikedArtworksConnection", () => {
+    const query = `
+    {
+      user(id: "blah") {
+        dislikedArtworksConnection(first: 10) {
+          totalCount
+          edges {
+            node {
+              title
+            }
+          }
+        }
+      }
+    }
+  `
+
+    const user = {
+      id: "blah",
+    }
+
+    const artworks = [
+      {
+        title: "Not A Fan",
+      },
+      {
+        title: "Don't Like This",
+      },
+    ]
+
+    let context
+
+    beforeEach(() => {
+      context = {
+        userByIDLoader: () => {
+          return Promise.resolve(user)
+        },
+        dislikedArtworksLoader: () => {
+          return Promise.resolve({
+            body: artworks,
+            headers: { "x-total-count": "2" },
+          })
+        },
+      }
+    })
+
+    it("returns disliked artworks for a user", async () => {
+      const {
+        user: {
+          dislikedArtworksConnection: { totalCount, edges },
+        },
+      } = await runAuthenticatedQuery(query, context)
+
+      expect(totalCount).toEqual(2)
+      expect(edges.length).toEqual(2)
+      expect(edges[0]).toEqual({
+        node: {
+          title: "Not A Fan",
+        },
+      })
+      expect(edges[1]).toEqual({
+        node: {
+          title: "Don't Like This",
+        },
+      })
+    })
+
+    it("returns an empty connection w/ no error if the gravity request 404's", async () => {
+      context.dislikedArtworksLoader = () => {
+        return Promise.reject(new HTTPError("Not Found", 404))
+      }
+
+      const {
+        user: {
+          dislikedArtworksConnection: { totalCount, edges },
+        },
+      } = await runAuthenticatedQuery(query, context)
+
+      expect(totalCount).toEqual(0)
+      expect(edges).toEqual([])
+    })
+
+    it("throws an error if the gravity request errors and it's not a 404", async () => {
+      context.dislikedArtworksLoader = () => {
+        return Promise.reject(new HTTPError("Cats in the server room", 500))
+      }
+
+      expect.assertions(1)
+
+      try {
+        await runAuthenticatedQuery(query, context)
+        throw new Error("An error was not thrown but was expected to throw.")
+      } catch (error) {
+        // eslint-disable-next-line jest/no-conditional-expect, jest/no-try-expect
+        expect(error.message).toEqual("Cats in the server room")
+      }
     })
   })
 })
